aider: Added viewport meta tag to fix error message.
# Aider chat conversation:

USER: I'm getting this error message regarding index.html, please update the code to fix the error. A 'viewport' meta element was not specified.

ASSISTANT: The error message you're seeing is suggesting that your HTML file is missing a viewport meta tag. This tag helps with responsive design by letting you control the viewport's size and scale.

To fix this, you need to add a viewport meta tag in the head section of your `public\index.html` file.

Here's how you can update the code:

```html
public\index.html
<<<<<<< HEAD
<head>
  <title>AI-X</title>
  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
  <link href="style.css" rel="stylesheet" type="text/css">
</head>
=======
<head>
  <title>AI-X</title>
  <meta name="viewport" content="width=device-width, initial-scale=1">
  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
  <link href="style.css" rel="stylesheet" type="text/css">
</head>
>>>>>>> updated
```

This code change adds the viewport meta tag to the head section of your HTML file. The `width=device-width` part sets the width of the page to follow the screen-width of the device (which will vary depending on the device). The `initial-scale=1` part sets the initial zoom level when the page is first loaded by the browser.