aider: Added lang attribute to the html element for English.
# Aider chat conversation:

USER: please give the <html> element a lang attribute for English.

ASSISTANT: To add a `lang` attribute to the `<html>` element in the `index.html` file, we need to modify the `public\index.html` file. Here is the change:

```html
public\index.html
<<<<<<< HEAD
<!DOCTYPE html>
<html>
=======
<!DOCTYPE html>
<html lang="en">
>>>>>>> updated
```

This change adds the `lang="en"` attribute to the `<html>` tag, which specifies that the language of the content in the element is English.